Sync proof of concept
It automatically schedules sync every ten seconds, and /schedule_sync
schedules a sync immediately.

The sync operation itself is a stub two-second sleep.

This is mostly a demonstration of how to fit multiple things in the one
process, and how to send messages between them. I look forward to an
opportunity to send channels down channels. That’s fun.

It *might* be worth putting sync onto a dedicated thread of its own.
As it stands it won’t matter much because the actix-web stuff creates
threads for itself, so sync and scheduler are all that’s left for the
main tokio… scheduler. Hmm, forgot they call it a scheduler now, used to
be called an executor. Ah well, two things called schedulers for now.

Fixes #207.