Work around what seems to be a rustc limitation
Where clause constraints on super associated types don’t seem to be
working properly in Rust. Can’t find anything in the rustc issue tracker
so I’ll minimise and report it because I believe it’s a bug.

So I threaded a duplicate associated type through, and that seems to
work around the problem.