Merge pool/connection macros
I’m of two minds about this. It *works*, but it’s not the most elegant.
I wouldn’t object to a vote to not include this.

As demonstrated, a drop-in replacement. Currently, anyway. The use of
closures can sometimes cause borrow checker problems that would not
otherwise occur, and it changes what return and ? mean, so it’s easy to
write code for which it wouldn’t be a drop-in replacement, but I don’t
think that’s an issue.

I contemplated making with_connection take a single closure and do an
enum<P, S> { Pg(P), Sqlite(S) }, which would allow you to avoid
duplicating the query code in the usual case, but it’s not worth the
slightly greater complexity because the only thing it’d help would be
generated code size when both databases are enabled, and the difference
is fairly negligible anyway; as far as runtime is concerned, there’ll be
no difference.

Next I’ll rename the macros at the call sites.