Brand postgres and sqlite queries in the macro
Putting it in the syntax is better than depending on comments.
I’ve used =>, but I think this may actually cause rustfmt to shut down
and not format the contents of the macro. Up to you if that matters or
not. It could be spelled = instead, that way it’d still parse as an
expression and rustfmt would I think then feel free to go wild once
again, but rustfmt in macros is fiddly. And I hate rustfmt anyway.
Anyway, = would be more open to misunderstanding, so => is better.

I also added support for an optional trailing comma, like normal
function calls can haave, and used it in the two call sites that I was
changing already.