Refactor sync translation error
A bit of a hybrid mixture of changes suggested by AE and
CM. Defines a single `SyncTranslationError` for all tables, with a
`table_name` field for logging. Tempted to do a bit of off-roading
to tidy up the constants, but will do that in a separate commit so it
can be easily reverted if it is too out-of-scope.