Add cynic to gql invoices query test
Refactors gql invoices query test to use cynic query builder. As our
gql types are undergoing a lot of churn at the moment, I went the
route of building queries manually rather than using the provided
derive macros. The documentation isn't complete (and has a few
errors), so it took a lot of trial-and-error and digging through
the source code to figure it out. Should be fairly simple to just
copy-and-paste the general pattern for other tests, though.